# Fix convergence for dolly+stage3 training