--- a/orttraining/orttraining/core/framework/torch/torch_proxy.cc
+++ b/orttraining/orttraining/core/framework/torch/torch_proxy.cc
@@ -10,9 +10,7 @@
 #include "orttraining/core/framework/torch/gil.h"
 #include "core/platform/env.h"
 
-namespace onnxruntime {
-namespace language_interop_ops {
-namespace torch {
+namespace onnxruntime::language_interop_ops::torch {
 
 void PythonObjectDeleter(PyObject* ptr) { Py_XDECREF(ptr); };
 
@@ -130,6 +128,18 @@ PyObject* CreateRequiresGradFlags(
   return flags;
 }
 
+PyObject* CreateInplaceMap(
+    const std::vector<int64_t>& inplace_map) {
+  PyObject* inplace_map_obj = Ort_PyList_New(inplace_map.size(), "inplace_map");
+
+  for (size_t output_index = 0; output_index < inplace_map.size(); ++output_index) {
+    PyObject* input_index = PyLong_FromLong(inplace_map[output_index]);
+    Ort_PyList_SetItem_NoIncref(inplace_map_obj, output_index, input_index, std::to_string(__LINE__));
+  }
+
+  return inplace_map_obj;
+}
+
 void InvokeRunner(
     PyObject* callback_runner,
     PyObject* args,
@@ -197,14 +207,15 @@ PythonObjectPtr CreatePythonCallArguments(
     const std::vector<void*>& obj_args,
     const std::vector<int64_t>& obj_indices,
     const bool is_training_mode,
-    const bool is_inplace,
-    const std::string& invoke_id) {
+    const std::vector<int64_t>& inplace_map,
+    const std::string& invoke_id,
+    const std::string& func_name) {
   ORT_ENFORCE(PyCallable_Check(callback), "Forward callback is not callable.");
   // The number of variables before those of
   // autograd.Function.apply and autograd.Function.backward.
   // The extra variables are used to configure the launch
   // forward and backward runners.
-  constexpr int64_t num_control_args = 6;
+  constexpr int64_t num_control_args = 7;
 
   // All arguments created for Python call will be destroyed along with PythonObjectPtr.
   PythonObjectPtr args(Ort_PyTuple_New(num_control_args + len, "forward_arguments_tuple"), PythonObjectDeleter);
@@ -216,10 +227,15 @@ PythonObjectPtr CreatePythonCallArguments(
   Ort_PyTuple_SetItem_NoIncref(args.get(), 2, tensor_flags, "tensor_flags");
   PyObject* is_training_mode_arg = is_training_mode ? Py_True : Py_False;
   Ort_PyTuple_SetItem_Incref(args.get(), 3, is_training_mode_arg, "is_training_mode");
-  PyObject* is_inplace_arg = is_inplace ? Py_True : Py_False;
-  Ort_PyTuple_SetItem_Incref(args.get(), 4, is_inplace_arg, "is_inplace_mode");
+
+  PyObject* inplace_map_arg = CreateInplaceMap(inplace_map);
+  Ort_PyTuple_SetItem_NoIncref(args.get(), 4, inplace_map_arg, "inplace_map");
+
   PyObject* kernel_invoke_id_arg = PyBytes_FromStringAndSize(invoke_id.c_str(), invoke_id.size());
   Ort_PyTuple_SetItem_NoIncref(args.get(), 5, kernel_invoke_id_arg, "kernel_invoke_id_arg");
+
+  PyObject* func_name_arg = PyBytes_FromStringAndSize(func_name.c_str(), func_name.size());
+  Ort_PyTuple_SetItem_NoIncref(args.get(), 6, func_name_arg, "func_name_arg");
 
   // Tensor inputs to call autograd.Function.apply or autograd.Function.backward.
   for (size_t i = 0; i < tensor_args.size(); ++i) {
@@ -246,6 +262,7 @@ PythonObjectPtr CreatePythonCallArguments(
 }
 
 void Invoke(
+    const std::string& func_name,
     PyObject* runner,
     PyObject* callback,
     const std::vector<int64_t>& requires_grads,
@@ -253,11 +270,11 @@ void Invoke(
     const std::vector<int64_t>& tensor_indices,
     const std::vector<void*>& obj_args,
     const std::vector<int64_t>& obj_indices,
-    void** diff_ctx,
-    std::vector<OrtValue>& returned_ortvalues,
     const bool is_training_mode,
-    const bool is_inplace,
-    const std::string& invoke_id) {
+    const std::vector<int64_t>& inplace_map,
+    const std::string& invoke_id,
+    void** diff_ctx,
+    std::vector<OrtValue>& returned_ortvalues) {
   const auto len = tensor_args.size() + obj_args.size();
   CheckArguments(len, requires_grads, tensor_args, tensor_indices, obj_args, obj_indices);
   RefCountTracker::GetInstance().Reset();
@@ -271,8 +288,9 @@ void Invoke(
         obj_args,
         obj_indices,
         is_training_mode,
-        is_inplace,
-        invoke_id);
+        inplace_map,
+        invoke_id,
+        func_name);
 
     RefCountTracker::GetInstance().DumpDetails("Before Invoke Python Call");
     InvokeRunner(runner, args.get(), is_training_mode, diff_ctx, returned_ortvalues);
@@ -282,17 +300,18 @@ void Invoke(
 }
 
 void TorchProxy::Forward(
+    const std::string& func_name,
     void* callback,
     const std::vector<int64_t>& requires_grads,
     const std::vector<std::optional<OrtValue>>& tensor_args,
     const std::vector<int64_t>& tensor_indices,
     const std::vector<void*>& obj_args,
     const std::vector<int64_t>& obj_indices,
-    void** diff_ctx,
-    std::vector<OrtValue>& returned_ortvalues,
     const bool is_training_mode,
-    const bool is_inplace,
-    const std::string& invoke_id) {
+    const std::vector<int64_t>& inplace_map,
+    const std::string& invoke_id,
+    void** diff_ctx,
+    std::vector<OrtValue>& returned_ortvalues) {
   // Semantically, this lock uniquely takes the ownership of TorchProxy
   // so that there will be only one of TorchProxy::Forward TorchProxy::Backward
   // can be run at one time.
@@ -301,6 +320,7 @@ void TorchProxy::Forward(
   GilGuard guard;
   auto runner = OrtTorchFunctionPool::GetInstance().GetForwardRunner();
   Invoke(
+      func_name,
       runner,
       reinterpret_cast<PyObject*>(callback),
       requires_grads,
@@ -308,22 +328,23 @@ void TorchProxy::Forward(
       tensor_indices,
       obj_args,
       obj_indices,
-      diff_ctx,
-      returned_ortvalues,
       is_training_mode,
-      is_inplace,
-      invoke_id);
+      inplace_map,
+      invoke_id,
+      diff_ctx,
+      returned_ortvalues);
 }
 
 void TorchProxy::Backward(
+    const std::string& func_name,
     void* callback,
     const std::vector<std::optional<OrtValue>>& tensor_args,
     const std::vector<int64_t>& tensor_indices,
     const std::vector<void*>& obj_args,
     const std::vector<int64_t>& obj_indices,
-    std::vector<OrtValue>& returned_ortvalues,
-    const bool is_inplace,
-    const std::string& invoke_id) {
+    const std::vector<int64_t>& inplace_map,
+    const std::string& invoke_id,
+    std::vector<OrtValue>& returned_ortvalues) {
   // Semantically, this lock uniquely takes the ownership of TorchProxy
   // so that there will be only one of TorchProxy::Forward TorchProxy::Backward
   // can be run at one time.
@@ -336,6 +357,7 @@ void TorchProxy::Backward(
   const auto all_input_count = tensor_args.size() + obj_args.size();
   const std::vector<int64_t> requires_grads(all_input_count, 0);
   Invoke(
+      func_name,
       runner,
       reinterpret_cast<PyObject*>(callback),
       requires_grads,
@@ -343,12 +365,11 @@ void TorchProxy::Backward(
       tensor_indices,
       obj_args,
       obj_indices,
-      nullptr /* context to store */,
-      returned_ortvalues,
       true /* is_training_mode */,
-      is_inplace,
-      invoke_id);
+      inplace_map,
+      invoke_id,
+      nullptr /* context to store */,
+      returned_ortvalues);
 }
-}  // namespace torch
-}  // namespace language_interop_ops
-}  // namespace onnxruntime
+
+}  // namespace onnxruntime::language_interop_ops::torch