--- a/orttraining/orttraining/test/python/orttraining_test_ortmodule_autograd.py
+++ b/orttraining/orttraining/test/python/orttraining_test_ortmodule_autograd.py
@@ -1549,7 +1549,7 @@ def test_python_op_save_input_for_backward():
                     count += 1
 
         if index == 0:
-            assert count == 1
+            assert count == 2
         else:
             assert count == 0